requirements.txt
datasets==2.14.5
faiss_gpu==1.7.2
huggingface_hub==0.22.2
langchain_community==0.2.4
langchain_core==0.2.5
numpy==1.26.4
openai==1.33.0
pandas==2.2.2
Requests==2.32.3
scikit_learn==1.4.0
sentence_transformers==2.3.1
torch==2.0.1
tqdm==4.66.1
transformers==4.37.2
